Update main.blocks, main.ts

+++ main.ts
@@ -9,17 +9,30 @@ input.onButtonPressed(Button.B, function () {
 })
 let turtle_rain: game.LedSprite = null
 let Player: game.LedSprite = null
+game.startCountdown(12000000)
 game.setScore(0)
-game.startCountdown(120000)
 Player = game.createSprite(2, 3)
 basic.forever(function () {
     turtle_rain = game.createSprite(randint(0, 4), 0)
-    for (let down = 0; down <= 4; down++) {
-        turtle_rain.change(LedSpriteProperty.Y, 1)
-        basic.pause(500)
-        if (turtle_rain.isTouching(Player)) {
-            game.addScore(1)
+    if (game.score() <= 10) {
+        for (let down = 0; down <= 4; down++) {
+            turtle_rain.change(LedSpriteProperty.Y, 1)
+            basic.pause(500)
+            if (turtle_rain.isTouching(Player)) {
+                game.addScore(1)
+            }
+        }
+        turtle_rain.delete()
+    } else {
+        if (game.score() > 10) {
+            for (let down = 0; down <= 4; down++) {
+                turtle_rain.change(LedSpriteProperty.Y, 1)
+                basic.pause(200)
+                if (turtle_rain.isTouching(Player)) {
+                    game.addScore(1)
+                }
+            }
+            turtle_rain.delete()
         }
     }
-    turtle_rain.delete()
 })
